Inline included source files in ecl_file.c
